requirements.txt
absl-py==1.4.0
albumentations==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.9.1
bleach==6.0.0
blessed==1.20.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.10.0
debugpy==1.6.8
decorator==5.1.1
defusedxml==0.7.1
easydict==1.12
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.18.0
fonttools==4.42.0
fqdn==1.5.1
fvcore==0.1.5.post20221221
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gpustat==1.1.1
grpcio==1.56.2
h5py==3.10.0
hydra-core==1.3.2
idna==2.10
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
iopath==0.1.9
ipykernel==6.25.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
lazy_loader==0.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
MedPy==0.4.0
mistune==3.0.1
ml-collections==0.1.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook==7.0.1
notebook_shim==0.2.3
numpy==1.24.4
nvidia-ml-py==12.535.133
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.8.0.74
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
portalocker==2.7.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydicom==2.4.2
Pygments==2.15.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
qudida==0.0.4
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.7
rpds-py==0.9.2
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
Send2Trash==1.8.2
SimpleITK==2.3.0
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
supervision==0.14.0
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.3.2
tinycss2==1.2.1
tomli==2.0.1
torch==1.7.1+cu110
torchvision==0.8.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
yacs==0.1.8
zipp==3.16.2
